requirements.txt
pandas~=2.0.3
python-dotenv~=1.0.1
Flask~=3.0.2
psycopg2-binary~=2.9.9
gunicorn~=21.2.0
Flask-Cors~=4.0.0
